Review Request 3202: Add field generation for datacomposition.json for the data flow.

+++ le-domain/src/test/java/com/latticeengines/domain/exposed/scoringapi/FieldTypeUnitTestNG.java
@@ -54,7 +54,7 @@ public class FieldTypeUnitTestNG {
         return new Object[][] { //
             { "boolean", FieldType.BOOLEAN }, //
             { "int", FieldType.INTEGER }, //
-            { "long", FieldType.INTEGER }, //
+            { "long", FieldType.LONG }, //
             { "float", FieldType.FLOAT }, //
             { "double", FieldType.FLOAT }, //
             { "string", FieldType.STRING } //
